Added encoding magic comment for Ruby 2.
Code93 has a full-ascii option which uses four control characters to
help with encoding.  I use high ascii (0x80 through 0x83) to represent
the control characters.  Since Ruby 2 has a default character encoding
of UTF8 these are no longer valid.  I force the encoding for that file
to be ASCII-8BIT which will cause it to simply work as in older Rubies.